Split misc browser journey
